pyproject.toml
[project]
name = "handffs-childcare"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.6.2",
    "google-auth>=2.40.3",
    "litellm>=1.74.7",
    "openai>=1.97.0",
    "openai-agents>=0.2.2",
    "py-dotenv>=0.1",
]
